pyproject.toml
[project]
name = "ui-automation"
version = "0.1.0"
description = "UI test automation with Selenium, Pytest, Allure"
authors = [{ name = "Dmitry Pakhomov", email = "you@example.com" }]
requires-python = ">=3.9"

dependencies = [
    "selenium>=4.11.2",
    "pytest>=7.4.0",
    "allure-pytest>=2.13.2"
]

[tool.pytest.ini_options]
addopts = "-v --tb=short"
markers = ["smoke: subset of smoke tests"]
